Remove PODObj.h
Replaces all users of PODObj with unique_ptr. This is problematic
because these users sometimes call Release() and the new users
later free the raw pointers with MemFree instead of delete. GCC/
glibc is ok with this for now but in the long run all uses of
MemFree must be removed.