Add smart pointer version of AddVideoObjectFromFile
Add this function to VObject.h to make it usable for the rest of
the project.

Use it in BobbyRUsed.cc for testing purposes. Pros: Makes that file
free of global variables. Cons: needs more CPU cycles to flip
through the pages of Bobby Ray's used items, shouldn't really be
a problem.