Mass dump of my current tree
Not going to have time to work on this in the foreseeable future
so I am dropping this now to keep it safe in case I want to come
back to it later. Most of the changes are cosmetic, replacing
old school C list and memory handling with C++14 constructs or
replacing copy & pasted sections with lambdas.