Refactor Smacker video code
Stop pretending that Smack_Stub.cc has anywhere near the functionality
of the original RAD Video Tools Smacker library. There is no need to
ape the original API with lots of constants and arguments that are not
used anywhere and functions that don't do anything. Instead this
commit strips it down to the minimum required to play our intro and
end game videos.

Additionally this commit
* fixes several memory leaks in Smack_Stub.cc, mostly only in case
of an error during SmackOpen.
* replaces the code specfic for the 565 pixel format with generic
SDL functions.
* makes the main window resizable.
* uses SDL_FULLSCREEN_DESKTOP instead of FULLSCREEN_DESKTOP which
makes switching from and to JA2 much more pleasant and doesn't
leave us with a 640x480 desktop in case of a crash.