Implement a new rgba32 type
This is a small wrapper around SDL_Color. The intended use of this type
is to replace the often seen combination 'Get16BPPColor(FROMRGB(...))'
and instead leave it to a few central places to convert from RGB32 to
RGB565.

As a demonstration this commit includes a new overload for LineDraw
which is then used to draw the bars in the soldier panels.