Add protective mouse regions over the screen when popups are active
This prevents the problem that the key ring or item stack popup may
stay active when the user clicks outside the area that the code tried
to restrict the cursor to via a call to the unimplemented function
RestrictMouseCursor().

Since this trick of using a do-nothing mouse region over the screen is
used in several places this commit factors it out into a new function
AddCoverRegion.