Remove GuiButtonRef
GuiButtonRef was an abstraction over raw GUI_BUTTON pointers:
an INT32 as the index into an array of pointers, ButtonList.
This was probably done to keep track of all currently active
buttons, which is now implemented with a std::set. This
removes the limit of 400 active buttons and also means we no
longer have to iterate over all 400 pointers, most of which
will be null, when we draw the buttons.