Lay the foundation for 32-bit blitters
This commit introduces the first of the new-style blitters.
These blitters are methods of a templated struct; the struct
can be used in either a 16-bit or a 32-bit version.

The blitters are instantiated for both 16-bit and 32-bit.

The first blitter replaces the old MonoShadowClip blitter,
the blitter used to print much of our text. This allows a
simple functionality of this approach in the FPS display,
which is the only part of the code base which uses its own
separate SDL texture.