Introduce the Persistant concept
This concept is used for types that must be read and written
member-by-member in specialized functions. With this commit
such types now implement templated Load/Save functions instead
of dedicated InjectTypeName/ExtractTypeName functions.

The advantage of this approach is that we just have to declare
the template function once and can remove header files like
LoadSaveRottingCorpse.h that just declare the two special
functions.