Use the new callback class to simplify existing code
This replaces the existing boilerplate classes which are
essentially just closures done the hard, pre C+11 way.

Also fixes a small bug in CharacterDialogueUsingAlternateFile
where we didn't pass on the correct quote. This affected only
the quest debug system; the only other user of this function
used the correct value already.