Reimplement 16 blitters
The old blitters were implemented with a mixture of copy
and paste and lots of individual changes. The new blitters
in this commit are all based on two common outer function
templates "Unrestricted" and "Clipped" which in their inner
loop call the blitter's 'core', the few lines of code that
actually write something to the destination buffer.